+++ demos/bookstore/app/router.ts
@@ -1,11 +1,11 @@
 import { createRouter } from '@remix-run/fetch-router'
-import { asyncContext } from '@remix-run/fetch-router/async-context-middleware'
+import { asyncContext } from '@remix-run/async-context-middleware'
 import { compression } from '@remix-run/fetch-router/compression-middleware'
-import { formData } from '@remix-run/fetch-router/form-data-middleware'
-import { logger } from '@remix-run/fetch-router/logger-middleware'
-import { methodOverride } from '@remix-run/fetch-router/method-override-middleware'
-import { session } from '@remix-run/fetch-router/session-middleware'
-import { staticFiles } from '@remix-run/fetch-router/static-middleware'
+import { formData } from '@remix-run/form-data-middleware'
+import { logger } from '@remix-run/logger-middleware'
+import { methodOverride } from '@remix-run/method-override-middleware'
+import { session } from '@remix-run/session-middleware'
+import { staticFiles } from '@remix-run/static-middleware'
 
 import { routes } from '../routes.ts'
 import { sessionCookie, sessionStorage } from './utils/session.ts'
@@ -33,7 +33,6 @@ middleware.push(
     cacheControl: 'no-store, must-revalidate',
     etag: false,
     lastModified: false,
-    acceptRanges: false,
   }),
 )
 middleware.push(formData({ uploadHandler }))