+++ packages/fetch-router/src/form-data-middleware.ts
@@ -1,4 +1,0 @@
-export { FormDataParseError } from '@remix-run/form-data-parser'
-export type { FileUpload, FileUploadHandler } from '@remix-run/form-data-parser'
-
-export { type FormDataOptions, formData } from './lib/middleware/form-data.ts'