+++ packages/logger-middleware/src/index.ts
@@ -1,0 +1,2 @@
+export { type LoggerOptions, logger } from './lib/logger.ts'
+